+++ src/core/processor/mod.rs
@@ -11,7 +11,8 @@ use std::{
 use super::{changes::Changes, queue::Queue, tree::Tree};
 use crate::{
 	argon_error,
-	constants::{BLACKLISTED_PATHS, CHANGES_THRESHOLD},
+	config::Config,
+	constants::BLACKLISTED_PATHS,
 	lock, logger,
 	project::{Project, ProjectDetails},
 	server, stats,
@@ -173,7 +174,7 @@ impl Handler {
 
 		trace!("Received client event: {:?} changes", changes.total());
 
-		if changes.total() > CHANGES_THRESHOLD {
+		if changes.total() > Config::new().changes_threshold {
 			let accept = logger::prompt(
 				&format!(
 					"You are about to apply {}, {} and {}. Do you want to continue?",